feat(ghost-v3): GHOST-006 sandbox dispatch + headless evidence capture baseline
- Add SandboxBackend trait to archon-permissions (dependency inversion)
- Wire sandbox flag through CommandContext, not private AtomicBool
- Gate both tool-execution paths (dispatch.rs + agent.rs) via sandbox check
- Implement SandboxBackend for SandboxGuard in archon-tui
- Headless evidence capture: 6 changes for integration test support
  (pub CtxBuilder, pub CommandContext, test-support feature, pub mod command)
- Integration test: tests/sandbox_dispatch_gate.rs (4 tests, both paths)
- cargo fmt --all across workspace
- Clippy fixes: collapsible_if → let-chains, await_holding_lock fixes,
  field_reassign_with_default, redundant pattern matching, etc.

+++ crates/archon-observability/src/metrics/channel.rs
@@ -162,7 +162,7 @@ impl ChannelMetrics {
             max_batch_size: self.max_batch_size.load(Ordering::Relaxed),
             p95_send_to_render_ms: {
                 let guard = self.p95_send_to_render_ms.lock();
-                guard.value_at_percentile(95.0) as u64
+                guard.value_at_percentile(95.0)
             },
         }
     }
